% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalMetrics.R
\name{calculateEvalMetrics}
\alias{calculateEvalMetrics}
\title{Calculate evaluation metrics of test mixed spot transcriptional profiles}
\usage{
calculateEvalMetrics(object, metrics = c("MAE", "MSE"))
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object with a trained model in
the \code{trained.model} slot and the actual cell proportions of test mixed
spot profiles in \code{prob.cell.types} slot.}

\item{metrics}{Metrics used to evaluate the model performance. Mean absolute
error (\code{"MAE"}) and mean squared error (\code{"MSE"}) by default.}
}
\value{
A \code{\linkS4class{SpatialDDLS}} object with the
\code{trained.model} slot containing a \code{\linkS4class{DeconvDLModel}}
object with \code{test.deconv.metrics} slot containing the calculated
metrics.
}
\description{
Calculate evaluation metrics of test mixed spot transcriptional profiles. By
default, absolute error (\code{AbsErr}), proportional absolute error
(\code{ppAbsErr}), squared error (\code{SqrErr}) and proportional squared
error (\code{ppSqrErr}) are calculated for each test mixed profile. In
addition, each of these metrics is aggregated according to three criteria:
cell type (\code{CellType}), probability bins in ranges of 0.1 (\code{pBin}),
and number of different cell types present in the spot \code{nCellTypes}.
}
\examples{
\dontrun{
set.seed(123)
sce <- SingleCellExperiment::SingleCellExperiment(
  assays = list(
    counts = matrix(
      rpois(30, lambda = 5), nrow = 15, ncol = 20,
      dimnames = list(paste0("Gene", seq(15)), paste0("RHC", seq(20)))
    )
  ),
  colData = data.frame(
    Cell_ID = paste0("RHC", seq(20)),
    Cell_Type = sample(x = paste0("CellType", seq(6)), size = 20,
                       replace = TRUE)
  ),
  rowData = data.frame(
    Gene_ID = paste0("Gene", seq(15))
  )
)
DDLS <- loadSCProfiles(
  single.cell.data = sce,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "Gene_ID"
)
probMatrixValid <- data.frame(
  Cell_Type = paste0("CellType", seq(6)),
  from = c(1, 1, 1, 15, 15, 30),
  to = c(15, 15, 30, 50, 50, 70)
)
DDLS <- generateBulkCellMatrix(
  object = DDLS,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_Type",
  prob.design = probMatrixValid,
  num.bulk.samples = 50,
  verbose = TRUE
)
# training of DDLS model
tensorflow::tf$compat$v1$disable_eager_execution()
DDLS <- trainSpatialDDLSModel(
  object = DDLS,
  on.the.fly = TRUE,
  batch.size = 15,
  num.epochs = 5
)
# evaluation using test data
DDLS <- calculateEvalMetrics(
  object = DDLS
)
}

}
\seealso{
\code{\link{distErrorPlot}} \code{\link{corrExpPredPlot}}
\code{\link{blandAltmanLehPlot}} \code{\link{barErrorPlot}}
}
