% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{cell.names}
\alias{cell.names}
\alias{cell.names,PropCellTypes-method}
\alias{cell.names<-}
\alias{cell.names<-,PropCellTypes-method}
\title{Get and set \code{cell.names} slot in a
\code{\linkS4class{PropCellTypes}} object}
\usage{
cell.names(object)

cell.names(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{Matrix containing names of the mixed transcriptional profiles to
be simulated as rows and cells to be used to simulate them as columns.}
}
\description{
Get and set \code{cell.names} slot in a
\code{\linkS4class{PropCellTypes}} object
}
