% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deconv.spots}
\alias{deconv.spots}
\alias{deconv.spots,SpatialDDLS-method}
\alias{deconv.spots<-}
\alias{deconv.spots<-,SpatialDDLS-method}
\title{Get and set \code{deconv.spots} slot in a
\code{\linkS4class{SpatialDDLS}} object}
\usage{
deconv.spots(object, index.st = NULL)

deconv.spots(object, index.st = NULL) <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{index.st}{Name or index of the results (same as for the
\code{spatial.experiments} slot). If \code{NULL} (by default), all results
are returned.}

\item{value}{List with deconvolution results of the data stored in the
\code{spatial.experiments} slot.}
}
\description{
Get and set \code{deconv.spots} slot in a
\code{\linkS4class{SpatialDDLS}} object
}
