% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnnModel.R
\name{deconvSpatialDDLS}
\alias{deconvSpatialDDLS}
\title{Deconvolute spatial transcriptomics mixed spots}
\usage{
deconvSpatialDDLS(
  object,
  index.st,
  batch.size = 64,
  normalize = TRUE,
  scaling = "standarize",
  simplify.set = NULL,
  simplify.majority = NULL,
  use.generator = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object with
\code{trained.model} and \code{spatial.experiments} slots.}

\item{index.st}{Name or index of the dataset/slide stored in the
\code{SpatialDDLS} object (\code{spatial.experiments} slot) to be
deconvolute. If missing, all datasets will be deconvoluted.}

\item{batch.size}{Number of samples per gradient update. If not specified,
\code{batch.size} will default to 128.}

\item{normalize}{Normalize data before deconvolution (\code{TRUE} by
default).}

\item{scaling}{How to scale data before training. It may be:
\code{"standarize"} (values are centered around the mean with a unit
standard deviation) or \code{"rescale"} (values are shifted and rescaled so
that they end up ranging between 0 and 1). If \code{normalize = FALSE},
data are not scaled.}

\item{simplify.set}{List specifying which cell types should be compressed
into a new label whose name will be the list item. See examples for
details. If provided, results are stored in a list with \code{'raw'} and
\code{'simpli.set'} elements.}

\item{simplify.majority}{List specifying which cell types should be
compressed into the cell type with the highest proportion in each spot.
Unlike \code{simplify.set}, no new labels are created. If provided, results
are stored in a list with \code{'raw'} and \code{'simpli.majority'}
elements.}

\item{use.generator}{Boolean indicating whether to use generators for
prediction (\code{FALSE} by default).}

\item{verbose}{Show informative messages during the execution.}
}
\value{
\code{\linkS4class{SpatialDDLS}} object with \code{deconv.spots}
slot. The output is a data frame with spot (\eqn{i}) as rows and cell types
(\eqn{j}) as columns. Each entry represents the proportion of \eqn{j} cell
type in \eqn{i} spot If \code{simplify.set} and/or
\code{simpplify.majority} are provided, the \code{deconv.spots} slot will
contain a list with raw and simplified results.
}
\description{
Deconvolute spatial transcriptomics mixed spots using the trained model
contained in the \code{\linkS4class{SpatialDDLS}} object.
}
\details{
This function requires a \code{SpatialDDLS} object with a trained Deep Neural
Network model (\code{\link{trained.model}} slot, see
\code{?\link{trainDeconvModel}}) and the spatial transcriptomics datasets to
be deconvoluted in the \code{spatial.experiments} slot. See
\code{?\link{createSpatialDDLSobject}} or \code{?\link{loadSTProfiles}} for
more details.
}
\examples{
\dontrun{
set.seed(123)
sce <- SingleCellExperiment::SingleCellExperiment(
  assays = list(
    counts = matrix(
      rpois(30, lambda = 5), nrow = 15, ncol = 20,
      dimnames = list(paste0("Gene", seq(15)), paste0("RHC", seq(20)))
    )
  ),
  colData = data.frame(
    Cell_ID = paste0("RHC", seq(20)),
    Cell_Type = sample(x = paste0("CellType", seq(6)), size = 20,
                       replace = TRUE)
  ),
  rowData = data.frame(
    Gene_ID = paste0("Gene", seq(15))
  )
)
DDLS <- loadSCProfiles(
  single.cell.data = sce,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "Gene_ID"
)
probMatrixValid <- data.frame(
  Cell_Type = paste0("CellType", seq(6)),
  from = c(1, 1, 1, 15, 15, 30),
  to = c(15, 15, 30, 50, 50, 70)
)
DDLS <- generateBulkCellMatrix(
  object = DDLS,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_Type",
  prob.design = probMatrixValid,
  num.bulk.samples = 50,
  verbose = TRUE
)
# training of DDLS model
tensorflow::tf$compat$v1$disable_eager_execution()
DDLS <- trainDeconvModel(
  object = DDLS,
  on.the.fly = TRUE,
  batch.size = 15,
  num.epochs = 5
)
# simulating bulk RNA-Seq data
countsBulk <- matrix(
  stats::rpois(100, lambda = sample(seq(4, 10), size = 100, replace = TRUE)),
  nrow = 40, ncol = 15,
  dimnames = list(paste0("Gene", seq(40)), paste0("Bulk", seq(15)))
)
seBulk <- SummarizedExperiment(assay = list(counts = countsBulk))
DDLS <- loadDeconvData(
  object = DDLS,
  data = seBulk,
  index.st = "Example"
)
# simplify arguments
simplify <- list(CellGroup1 = c("CellType1", "CellType2", "CellType4"),
                 CellGroup2 = c("CellType3", "CellType5"))
DDLS <- deconvSpatialDDLS(
  object = DDLS,
  index.st = "Example",
  simplify.set = simplify,
  simplify.majority = simplify
)
}
  
}
\seealso{
\code{\link{trainDeconvModel}} \code{\linkS4class{SpatialDDLS}}
}
