% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,DeconvDLModel-method}
\alias{features<-}
\alias{features<-,DeconvDLModel-method}
\title{Get and set \code{features} slot in a
\code{\linkS4class{DeconvDLModel}} object}
\usage{
features(object)

features(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{value}{Vector with features (genes) considered by the deep neural
network model.}
}
\description{
Get and set \code{features} slot in a
\code{\linkS4class{DeconvDLModel}} object
}
