% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getProbMatrix}
\alias{getProbMatrix}
\title{Getter function for the cell composition matrix}
\usage{
getProbMatrix(object, type.data)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object with
\code{prob.cell.types} slot.}

\item{type.data}{Subset of data to return: \code{train} or \code{test}.}
}
\value{
Cell type proportion matrix.
}
\description{
Getter function for the cell composition matrix. This function allows to
access to the cell composition matrix of simulated mixed transcriptional
profiles.
}
\seealso{
\code{\link{genMixedCellProp}}
}
