% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToDDLS}
\alias{listToDDLS}
\title{Transform SpatialDDLS-like list into an actual SpatialDDLSDNN object}
\usage{
listToDDLS(listTo)
}
\arguments{
\item{listTo}{A list in which each element must correspond to each slot of an
\code{SpatialDDLS} object. The names must be the same as the slot
names.}
}
\value{
\code{SpatialDDLS} object the data provided in the original list.
}
\description{
Transform SpatialDDLS-like list into an actual \code{SpatialDDLS}
object. This function allows to generate the examples and the vignettes of
\pkg{SpatialDDLS} package as these are the data used. These data are
stored in the SpatialDDLSdata package.
}
\seealso{
\code{\link{listToDDLSDNN}}
}
