% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listToDDLSDNN}
\alias{listToDDLSDNN}
\title{Transform SpatialDDLSDNN-like list into an actual SpatialDDLSDNN
object}
\usage{
listToDDLSDNN(listTo)
}
\arguments{
\item{listTo}{A list in which each element must correspond to each slot of an
\code{SpatialDDLSDNN} object. The names must be the same as the slot
names.}
}
\value{
\code{SpatialDDLSDNN} object with the data provided in the
original list.
}
\description{
Transform SpatialDDLSDNN-like list into an actual
\code{SpatialDDLSDNN} object. This function allows to use pre-trained
models in the \pkg{SpatialDDLS} package. These models are stored in the
SpatialDDLSmodels package.
}
\seealso{
\code{\link{listToDDLS}}
}
