% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{method}
\alias{method}
\alias{method,PropCellTypes-method}
\alias{method<-}
\alias{method<-,PropCellTypes-method}
\title{Get and set \code{method} slot in a \code{\linkS4class{PropCellTypes}}
object}
\usage{
method(object)

method(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{Vector containing the method by which cell type proportions were
generated.}
}
\description{
Get and set \code{method} slot in a \code{\linkS4class{PropCellTypes}}
object
}
