% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{mixed.spot.profiles}
\alias{mixed.spot.profiles}
\alias{mixed.spot.profiles,SpatialDDLS-method}
\alias{mixed.spot.profiles<-}
\alias{mixed.spot.profiles<-,SpatialDDLS-method}
\title{Get and set \code{mixed.spot.profiles} slot in a
\code{\linkS4class{SpatialDDLS}} object}
\usage{
mixed.spot.profiles(object, type.data = "both")

mixed.spot.profiles(object, type.data = "both") <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{type.data}{Element of the list. Can be \code{'train'}, \code{'test'} or
\code{'both'} (the last by default).}

\item{value}{List with two elements, train and test, each one being
a \code{\linkS4class{SummarizedExperiment}} object with simulated bulk
RNA-Seq samples.}
}
\description{
Get and set \code{mixed.spot.profiles} slot in a
\code{\linkS4class{SpatialDDLS}} object
}
