% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{model}
\alias{model}
\alias{model,DeconvDLModel-method}
\alias{model<-}
\alias{model<-,DeconvDLModel-method}
\title{Get and set \code{model} slot in a \code{\linkS4class{DeconvDLModel}}
object}
\usage{
model(object)

model(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{value}{\code{keras.engine.sequential.Sequential} object with a trained
deep neural network model.}
}
\description{
Get and set \code{model} slot in a \code{\linkS4class{DeconvDLModel}}
object
}
