% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plots}
\alias{plots}
\alias{plots,PropCellTypes-method}
\alias{plots<-}
\alias{plots<-,PropCellTypes-method}
\title{Get and set \code{plots} slot in a \code{\linkS4class{PropCellTypes}}
object}
\usage{
plots(object)

plots(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{List of lists with plots showing the distribution of cell
proportions generated by each method.}
}
\description{
Get and set \code{plots} slot in a \code{\linkS4class{PropCellTypes}}
object
}
