% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialCoor.R
\name{plotSpatialClustering}
\alias{plotSpatialClustering}
\title{Plot spatial clustering based on predicted cell proportions}
\usage{
plotSpatialClustering(
  object,
  index.st,
  method,
  k.nn,
  k.centers,
  colors,
  size.point = 1,
  title = NULL,
  theme = NULL
)
}
\arguments{
\item{object}{A \code{\linkS4class{SpatialDDLS}} object.}

\item{index.st}{Index of the spatial transcriptomics data to be plotted. It
can be either a position or a name if a named list was provided.}

\item{method}{Cell type predicted proportions to color spots by.}

\item{k.nn}{Color scale to be used. It can be \code{"blues"} or
\code{"spectral"} (the former by default).}

\item{k.centers}{If results were simplified (see \code{?\link{deconvSpatialDDLS}}
for details), what results to plot (\code{raw} by default).}

\item{colors}{A vector of two elements indicating wanted limits for color
scale. If \code{NULL} (by default), color scale is adjusted to max and min
predicted proportions.}

\item{size.point}{Size of points (0.1 by default).}

\item{title}{Title of plot.}

\item{theme}{\pkg{ggplot2} theme.}
}
\value{
A ggplot object.
}
\description{
Color spots on the spatial coordinates according to clustering basedd on
predicted proportions.
}
\seealso{
\code{\link{plotSpatialPropAll}} \code{\link{deconvSpatialDDLS}}
\code{\link{trainDeconvModel}}
}
