% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialCoor.R
\name{plotSpatialProp}
\alias{plotSpatialProp}
\title{Plot predicted proportions for a specific cell type using spatial coordinates
of spots}
\usage{
plotSpatialProp(
  object,
  index.st,
  cell.type,
  colors = NULL,
  set = "raw",
  size.point = 1,
  alpha.point = 1,
  title = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{index.st}{Nmae or index of the desired
\code{\linkS4class{SpatialExperiment}} object to visualize.}

\item{cell.type}{Cell type predicted proportions to color spots by.}

\item{size.point}{Size of points (0.1 by default).}

\item{alpha.point}{Alpha of points (0.1 by default).}

\item{title}{Title of plot.}

\item{theme}{\pkg{ggplot2} theme.}
}
\value{
A ggplot object.
}
\description{
Color spots on the spatial coordinates plot according to predicted
proportions of a particular cell type. Color scale is adapted depending on
the range of predicted proportions.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{plotSpatialPropAll}} \code{\link{deconvSpatialDDLS}}
\code{\link{trainDeconvModel}}
}
