% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialCoor.R
\name{plotSpatialPropAll}
\alias{plotSpatialPropAll}
\title{Plot predicted proportions for all cell types using spatial coordinates of
spots}
\usage{
plotSpatialPropAll(
  object,
  index.st,
  colors = NULL,
  set = "raw",
  size.point = 0.1,
  alpha.point = 1,
  title = NULL,
  nrow = NULL,
  ncol = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{colors}{Vector of colors to be used. Only vectors with a number of
colors equal to or greater than the levels of \code{color.by} will be
accepted. By default, a custom color list is used.}

\item{size.point}{Size of points (0.1 by default).}

\item{alpha.point}{Alpha of points (0.1 by default).}

\item{title}{Boolean indicating whether single-cell profiles are filtered out
and only errors associated with pseudo-bulk samples are displayed
(\code{TRUE} by default).}

\item{nrow}{Number of rows in the split plot.}

\item{ncol}{Nomber of columns in the split plot}

\item{theme}{\pkg{ggplot2} theme.}

\item{color.scale}{Variable used to display data in different panels. If
\code{NULL}, the plot is not split into different panels. Options are
\code{nCellTypes} (by number of different cell types) and \code{CellType}
(by cell type).}
}
\value{
A ggplot object.
}
\description{
Color spots on the spatial coordinates plot according to their predicted cell
type proportions. All cell types are represented altogether using the same
scale of colors (from 0 to 1).
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{plotSpatialProp}} \code{\link{deconvSpatialDDLS}}
\code{\link{trainDeconvModel}}
}
