% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{prob.cell.types}
\alias{prob.cell.types}
\alias{prob.cell.types,SpatialDDLS-method}
\alias{prob.cell.types<-}
\alias{prob.cell.types<-,SpatialDDLS-method}
\title{Get and set \code{prob.cell.types} slot in a
\code{\linkS4class{SpatialDDLS}} object}
\usage{
prob.cell.types(object, type.data = "both")

prob.cell.types(object, type.data = "both") <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{type.data}{Type of data to return. It can be \code{'train'},
\code{'test'} or \code{'both'} (the last by default).}

\item{value}{List with two elements corresponding to train and test, each one
with a \code{\linkS4class{PropCellTypes}} object.}
}
\description{
Get and set \code{prob.cell.types} slot in a
\code{\linkS4class{SpatialDDLS}} object
}
