% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{prob.matrix}
\alias{prob.matrix}
\alias{prob.matrix,PropCellTypes-method}
\alias{prob.matrix<-}
\alias{prob.matrix<-,PropCellTypes-method}
\title{Get and set \code{prob.matrix} slot in a
\code{\linkS4class{PropCellTypes}} object}
\usage{
prob.matrix(object)

prob.matrix(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{Matrix with cell types as columns and samples as rows.}
}
\description{
Get and set \code{prob.matrix} slot in a
\code{\linkS4class{PropCellTypes}} object
}
