% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{project}
\alias{project}
\alias{project,SpatialDDLS-method}
\alias{project<-}
\alias{project<-,SpatialDDLS-method}
\title{Get and set \code{project} slot in a \code{\linkS4class{SpatialDDLS}}
object}
\usage{
project(object)

project(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{value}{Character indicating the name of the project.}
}
\description{
Get and set \code{project} slot in a \code{\linkS4class{SpatialDDLS}}
object
}
