% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{set.list}
\alias{set.list}
\alias{set.list,PropCellTypes-method}
\alias{set.list<-}
\alias{set.list<-,PropCellTypes-method}
\title{Get and set \code{set.list} slot in a
\code{\linkS4class{PropCellTypes}} object}
\usage{
set.list(object)

set.list(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{List of cells sorted according to the cell type to which they
belong.}
}
\description{
Get and set \code{set.list} slot in a
\code{\linkS4class{PropCellTypes}} object
}
