% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{set}
\alias{set}
\alias{set,PropCellTypes-method}
\alias{set<-}
\alias{set<-,PropCellTypes-method}
\title{Get and set \code{set} slot in a \code{\linkS4class{PropCellTypes}}
object}
\usage{
set(object)

set(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{PropCellTypes}} object.}

\item{value}{A vector containing the names of cells that are present in the
object.}
}
\description{
Get and set \code{set} slot in a \code{\linkS4class{PropCellTypes}}
object
}
