% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{single.cell.real}
\alias{single.cell.real}
\alias{single.cell.real,SpatialDDLS-method}
\alias{single.cell.real<-}
\alias{single.cell.real<-,SpatialDDLS-method}
\title{Get and set \code{single.cell.real} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
single.cell.real(object)

single.cell.real(object) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{value}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
with real single-cell profiles.}
}
\description{
Get and set \code{single.cell.real} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
