% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{single.cell.simul}
\alias{single.cell.simul}
\alias{single.cell.simul,SpatialDDLS-method}
\alias{single.cell.simul<-}
\alias{single.cell.simul<-,SpatialDDLS-method}
\title{Get and set \code{single.cell.simul} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
single.cell.simul(object)

single.cell.simul(object) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{value}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
with simulated single-cell profiles.}
}
\description{
Get and set \code{single.cell.simul} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
