% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{spatial.experiments}
\alias{spatial.experiments}
\alias{spatial.experiments,SpatialDDLS-method}
\alias{spatial.experiments<-}
\alias{spatial.experiments<-,SpatialDDLS-method}
\title{Get and set \code{spatial.experiments} slot in a
\code{\linkS4class{SpatialDDLS}} object}
\usage{
spatial.experiments(object, name.data = NULL)

spatial.experiments(object, name.data = NULL) <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{name.data}{Name of the ST data. If \code{NULL} (by default), all
data contained in the \code{spatial.experiments} slot are returned.}

\item{value}{List whose names are the reference of the stored data.}
}
\description{
Get and set \code{spatial.experiments} slot in a
\code{\linkS4class{SpatialDDLS}} object
}
