% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDDLS.R
\docType{package}
\name{SpatialDDLS-package}
\alias{SpatialDDLS-package}
\title{SpatialDDLS: Deconvolution of Spatial Transcriptomics Data Based on Neural Networks}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Deconvolution of spatial transcriptomics data based on neural networks and single-cell RNA-seq data. SpatialDDLS implements a workflow to create neural network models able to make accurate estimates of cell composition of spots from spatial transcriptomics data using deep learning and the meaningful information provided by single-cell RNA-seq data. See Torroja and Sanchez-Cabo (2019) \doi{10.3389/fgene.2019.00978} and Mañanes et al. (2024) \doi{10.1093/bioinformatics/btae072} to get an overview of the method and see some examples of its performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://diegommcc.github.io/SpatialDDLS/}
  \item \url{https://github.com/diegommcc/SpatialDDLS}
  \item Report bugs at \url{https://github.com/diegommcc/SpatialDDLS/issues}
}

}
\author{
\strong{Maintainer}: Diego Mañanes \email{dmananesc@cnic.es} (\href{https://orcid.org/0000-0001-7247-6794}{ORCID})

Authors:
\itemize{
  \item Carlos Torroja \email{ctorroja@cnic.es} (\href{https://orcid.org/0000-0001-8914-3400}{ORCID})
  \item Fatima Sanchez-Cabo \email{fscabo@cnic.es} (\href{https://orcid.org/0000-0003-1881-1664}{ORCID})
}

}
\keyword{internal}
