% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDDLS.R
\docType{package}
\name{SpatialDDLS}
\alias{SpatialDDLS}
\title{SpatialDDLS: an R package to deconvolute spatial transcriptomics data using
single-cell RNA-Seq data and Neural Networks}
\description{
\pkg{SpatialDDLS} is an R package that allows to deconvolute spatial
transcriptomics data using context-specific deconvolution models based on
single-cell RNA-Seq data and Neural Networks. These models are able to make
accurate estimates of the cell composition of spots using the advances
provided by Deep Learning and the meaningful information provided by
scRNA-Seq data. See Torroja and Sanchez-Cabo (2019)
(\doi{10.3389/fgene.2019.00978}) for more details.
}
\details{
The foundation of the method consists of a process that starts from
single-cell RNA-Seq data and, after a few steps, a Deep Neural Network (DNN)
model is trained with simulated mixed spot transcriptional profiles whose cell
composition is known. See vignettes and
\url{https://diegommcc.github.io/SpatialDDLS/} for more details.
}
