% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{test.deconv.metrics}
\alias{test.deconv.metrics}
\alias{test.deconv.metrics,DeconvDLModel-method}
\alias{test.deconv.metrics<-}
\alias{test.deconv.metrics<-,DeconvDLModel-method}
\title{Get and set \code{test.deconv.metrics} slot in a
\code{\linkS4class{DeconvDLModel}} object}
\usage{
test.deconv.metrics(object, metrics = "All")

test.deconv.metrics(object, metrics = "All") <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{metrics}{Metrics to show (\code{'All'} by default)}

\item{value}{List with evaluation metrics to assess the performance of the
model on each sample of test data.}
}
\description{
Get and set \code{test.deconv.metrics} slot in a
\code{\linkS4class{DeconvDLModel}} object
}
