% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{test.metrics}
\alias{test.metrics}
\alias{test.metrics,DeconvDLModel-method}
\alias{test.metrics<-}
\alias{test.metrics<-,DeconvDLModel-method}
\title{Get and set \code{test.metrics} slot in a
\code{\linkS4class{DeconvDLModel}} object}
\usage{
test.metrics(object)

test.metrics(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{value}{List with evaluation metrics after prediction on test data.}
}
\description{
Get and set \code{test.metrics} slot in a
\code{\linkS4class{DeconvDLModel}} object
}
