% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{test.pred}
\alias{test.pred}
\alias{test.pred,DeconvDLModel-method}
\alias{test.pred<-}
\alias{test.pred<-,DeconvDLModel-method}
\title{Get and set \code{test.pred} slot in a
\code{\linkS4class{DeconvDLModel}} object}
\usage{
test.pred(object)

test.pred(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{value}{Matrix object with prediction results on test data.}
}
\description{
Get and set \code{test.pred} slot in a
\code{\linkS4class{DeconvDLModel}} object
}
