% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interGradientsDL.R
\name{topGradientsCellType}
\alias{topGradientsCellType}
\title{Get top genes with largest/smallest gradients per cell type}
\usage{
topGradientsCellType(object, method = "class", top.n.genes = 15)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object with a trained
deconvolution model (\code{trained.model} slot) and pure mixed
transcriptional profiles (\code{mixed.profiles} slot).}

\item{method}{Method to calculate gradients with respect to inputs. It can be
\code{'class'} (gradients of predicted classes w.r.t. inputs),
\code{'loss'} (gradients of loss w.r.t. inputs) or \code{'both'}.
\code{'class'} by default.}

\item{top.n}{Top n genes (positive and negative) taken per cell type.}
}
\value{
Object containing gradients in the \code{interpret.gradients} slot of
the \code{DeconvDLModel} object (\code{trained.model} slot).
}
\description{
Get feature names of top features with largest/smallest gradients per cell type.
These genes can be used as input to visualize how they are spatially expressed
(\code{plotGeneSpatial} function) or to plot gradients a a heatmap (\code{plotGradHeatmap} function).
}
\examples{
\donttest{
set.seed(123)
sce <- SingleCellExperiment::SingleCellExperiment(
  assays = list(
    counts = matrix(
      rpois(30, lambda = 5), nrow = 15, ncol = 10,
      dimnames = list(paste0("Gene", seq(15)), paste0("RHC", seq(10)))
    )
  ),
  colData = data.frame(
    Cell_ID = paste0("RHC", seq(10)),
    Cell_Type = sample(x = paste0("CellType", seq(2)), size = 10,
                       replace = TRUE)
  ),
  rowData = data.frame(
    Gene_ID = paste0("Gene", seq(15))
  )
)
SDDLS <- createSpatialDDLSobject(
  sc.data = sce,
  sc.cell.ID.column = "Cell_ID",
  sc.gene.ID.column = "Gene_ID",
)
SDDLS <- genMixedCellProp(
  object = SDDLS,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_Type",
  num.sim.spots = 50,
  verbose = TRUE
)
SDDLS <- simMixedProfiles(SDDLS)
SDDLS <- trainDeconvModel(
  object = SDDLS,
  batch.size = 12,
  num.epochs = 5
)
## calculating gradients
SDDLS <- interGradientsDL(SDDLS)
listGradients <- topGradientsCellType(SDDLS)
lapply(listGradients, head, n = 5)
}
  
}
\seealso{
\code{\link{interGradientsDL}} \code{\link{trainDeconvModel}}
}
