% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{trained.model}
\alias{trained.model}
\alias{trained.model,SpatialDDLS-method}
\alias{trained.model<-}
\alias{trained.model<-,SpatialDDLS-method}
\title{Get and set \code{trained.model} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
trained.model(object)

trained.model(object) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{value}{\code{\linkS4class{DeconvDLModel}} object.}
}
\description{
Get and set \code{trained.model} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
