% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{training.history}
\alias{training.history}
\alias{training.history,DeconvDLModel-method}
\alias{training.history<-}
\alias{training.history<-,DeconvDLModel-method}
\title{Get and set \code{training.history} slot in a
\code{\linkS4class{DeconvDLModel}} object}
\usage{
training.history(object)

training.history(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{DeconvDLModel}} object.}

\item{value}{\code{keras_training_history} object with the training history
of the deep neural network model.}
}
\description{
Get and set \code{training.history} slot in a
\code{\linkS4class{DeconvDLModel}} object
}
