% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{zinb.params}
\alias{zinb.params}
\alias{zinb.params,SpatialDDLS-method}
\alias{zinb.params<-}
\alias{zinb.params<-,SpatialDDLS-method}
\title{Get and set \code{zinb.params} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
zinb.params(object)

zinb.params(object) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{value}{\code{\linkS4class{ZinbParametersModel}} object with a valid
\code{\link[zinbwave]{zinbModel}} object.}
}
\description{
Get and set \code{zinb.params} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
