% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{zinbwave.model}
\alias{zinbwave.model}
\alias{zinbwave.model,ZinbParametersModel-method}
\alias{zinbwave.model<-}
\alias{zinbwave.model<-,ZinbParametersModel-method}
\title{Get and set \code{zinbwave.model} slot in a
\code{\linkS4class{ZinbParametersModel}} object}
\usage{
zinbwave.model(object)

zinbwave.model(object) <- value
}
\arguments{
\item{object}{\code{\linkS4class{ZinbParametersModel}} object.}

\item{value}{\code{\link[zinbwave]{zinbModel}} object with the
estimated parameters to simulate new single-cell profiles.}
}
\description{
Get and set \code{zinbwave.model} slot in a
\code{\linkS4class{ZinbParametersModel}} object
}
