% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deconv.spots}
\alias{deconv.spots}
\alias{deconv.spots,SpatialDDLS-method}
\alias{deconv.spots<-}
\alias{deconv.spots<-,SpatialDDLS-method}
\title{Get and set \code{deconv.spots} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
deconv.spots(object, index.st = NULL)

deconv.spots(object, index.st = NULL) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{index.st}{Name or index of predicted cell proportions (same as for the
\code{spatial.experiments} slot). If \code{NULL} (by default), all
results are returned.}

\item{value}{List of predicted cell type proportions for the experiments
stored in the \code{spatial.experiments} slot.}
}
\description{
Get and set \code{deconv.spots} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
