% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deconv.spots}
\alias{deconv.spots}
\alias{deconv.spots,SpatialDDLS-method}
\alias{deconv.spots<-}
\alias{deconv.spots<-,SpatialDDLS-method}
\title{Get and set \code{deconv.spots} slot in a
\code{\linkS4class{SpatialDDLS}} object}
\usage{
deconv.spots(object, name.data = NULL)

deconv.spots(object, name.data = NULL) <- value
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object.}

\item{name.data}{Name of the data. If \code{NULL} (by default), all
results contained in the \code{deconv.spots} slot are returned.}

\item{value}{List whose names are the reference of the stored results.}
}
\description{
Get and set \code{deconv.spots} slot in a
\code{\linkS4class{SpatialDDLS}} object
}
