% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotTrainingHistory}
\alias{plotTrainingHistory}
\title{Plot training history of a trained SpatialDDLS deep neural network model}
\usage{
plotTrainingHistory(
  object,
  title = "History of metrics during training",
  metrics = NULL
)
}
\arguments{
\item{object}{\code{\linkS4class{SpatialDDLS}} object with a
\code{trained.model} slot.}

\item{title}{Title of plot.}

\item{metrics}{Metrics to be plotted. If \code{NULL} (by default), all
metrics available in the \code{\linkS4class{DeconvDLModel}} object will be
plotted.}
}
\value{
A ggplot object with the progression of the selected metrics during
training.
}
\description{
Plot training history of a trained SpatialDDLS deep neural network model.
}
\seealso{
\code{\link{trainDeconvModel}}
}
