% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{prob.cell.types}
\alias{prob.cell.types}
\alias{prob.cell.types,SpatialDDLS-method}
\alias{prob.cell.types<-}
\alias{prob.cell.types<-,SpatialDDLS-method}
\title{Get and set \code{prob.cell.types} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
prob.cell.types(object, type.data = "both")

prob.cell.types(object, type.data = "both") <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{type.data}{Type of data to return. It can be \code{'both'} (default),
\code{'train'}, or \code{'test'}.}

\item{value}{List with two \code{\linkS4class{PropCellTypes}} objects
corresponding to train and test data.}
}
\description{
Get and set \code{prob.cell.types} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
