% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{spatial.experiments}
\alias{spatial.experiments}
\alias{spatial.experiments,SpatialDDLS-method}
\alias{spatial.experiments<-}
\alias{spatial.experiments<-,SpatialDDLS-method}
\title{Get and set \code{spatial.experiments} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object}
\usage{
spatial.experiments(object, index.st = NULL)

spatial.experiments(object, index.st = NULL) <- value
}
\arguments{
\item{object}{\code{\link[SpatialExperiment]{SpatialExperiment}} object.}

\item{index.st}{Index of the spatial transcriptomics data within the list. It
can be either an position or a name if a named list was provided. If
\code{NULL} (by default), all data contained in the
\code{spatial.experiments} slot are returned.}

\item{value}{List in which each element is a
\code{\link[SpatialExperiment]{SpatialExperiment}}
object. It can be a named list.}
}
\description{
Get and set \code{spatial.experiments} slot in a
\code{\link[SpatialExperiment]{SpatialExperiment}} object
}
